+++ lib/ruby_lokalise_api/resources/user.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module RubyLokaliseApi
+  module Resources
+    class User < Base
+      MAIN_PARAMS = [].freeze
+      no_support_for %i[update destroy reload_data]
+    end
+  end
+end